+++ CHANGELOG.md
@@ -13,13 +13,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 -   Cloud only: new regions: sa-bogota-1, us-dcc-swjordan-1,
     us-dcc-swjordan-2, ap-dcc-tokyo-1, me-abudhabi-3,
-    ap-dcc-gazipur-1, me-dcc-doha-1
+    ap-dcc-gazipur-1, me-dcc-doha-1, ap-hobsonville-1
 -   added support for array_collect() and count(distinct) in queries. These
     require server side support which is only available in Oracle NoSQL releases
     23.3 and higher and will not be immediately available in the cloud service
 -   Cloud only:  added support for Cloud feature Global Active Tables. This
     includes new requests and classes:
-    
+
     - AddReplicaRequest
     - DropReplicaRequest
     - ReplicaStatsRequest/Result
@@ -29,6 +29,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
     as well as additional replica-related information
     and interfaces in TableResult
 
+## Changed
+
+- Modified internal query processing to better support elasticity operations
 
 ## Fixed